build(esp32): add -Wconversion and re-enable IDF-suppressed warnings on first-party components
Second half of the follow-on from the host-test warning commit. ESP-IDF
already builds everything -Wall -Wextra, so main, ui and shared get
-Wconversion plus -Wunused-parameter and -Wsign-compare, which IDF turns
off globally and a later flag re-enables - each with an explicit
-Wno-error=, because IDF's -Werror=all escalates re-enabled -Wall-family
warnings and the policy here, as in the host suites, is warnings, not
-Werror. LVGL's interface include dirs are re-declared SYSTEM in the ui
and main components: lv_draw_buf.h's inline helpers alone fired 56
warnings in our TUs. shared was already clean; the rest sat almost
entirely in ui. Zero first-party warnings now, verified on a from-scratch
build (rm -rf build) through the pre-commit hook's own `make esp32` path
- the incremental builds along the way each hid a few TUs.

What the sweep found:

- The sample-load request's sample_rate wire field is a uint16_t hint, so
  a 96 kHz file's rate silently wrapped to 30464 on the wire. Daisy
  re-reads the real rate from the file, which is why nobody noticed. It
  now degrades to 0 = unknown instead of a wrong-but-plausible value.
- The UART TX wake marker was 0x7F cast into uart_event_type_t - outside
  the enum's value range, where the cast's result is formally unspecified.
  Now UART_EVENT_MAX: a real enumerator, well-defined, never posted by
  the driver.
- ~60 of the warnings had one root cause: Softkey::why was added without
  a default member initializer, so every {label, cb} aggregate init
  warned. One {} fixes them all; behavior is unchanged. Another 18 came
  from one unnamed-parameter fix in ui_page.h's default onInput.
- SampleBrowserState::isValid() tested selected_file_index >= 0 on an
  unsigned field - vacuously true since the sentinel it implies does not
  exist (0, the first entry, is a valid selection). The clause is gone;
  behavior is identical.
- Dead code: the unused play-index browser helper and its loop-gap
  constant, the computed-but-never-read directory_actually_changed, and
  two TAG variables with no remaining log calls.
- Mechanical: explicit int-to-float casts in display formatting and CV
  calibration scaling, uint8_t casts where the modulo or the protocol
  already bounds the value, int32_t for LVGL coordinate locals, size_t
  loop indices in pcnt_task, the uart_config_t/esp_timer_create_args_t
  fields spelled out, and casts on the LV_PART_x | LV_STATE_y enum ORs
  that C++2x deprecates.

Also fixes the check-build-esp32 hook itself: git commit exports a
relative GIT_INDEX_FILE (.git/index) to hooks, and the IDF Component
Manager's `git --work-tree /tmp/<tmp>` checkout resolved it inside the
temp work-tree, so the ESP32 build could never pass as a pre-commit hook
once a reconfigure was needed. `make esp32` now strips commit-time git
env before invoking idf.py.

+++ firmware/esp32/components/ui/pages/ui_cv_cal_page.cpp
@@ -54,13 +54,13 @@ std::atomic<bool> s_pending_flag{false};
 WaveX::Protocol::CvCalMessage ModelToMessage(uint8_t persist) {
     return WaveX::Protocol::CvCalMessage(static_cast<uint8_t>(s_model.group),
                                          persist,
-                                         s_model.cut_gain / 1000.0f,
-                                         s_model.cut_off / 1000.0f,
-                                         s_model.q_gain / 1000.0f,
-                                         s_model.q_off / 1000.0f,
-                                         s_model.vca_gain / 1000.0f,
-                                         s_model.vca_off / 1000.0f,
-                                         s_model.curve_k / 100.0f);
+                                         static_cast<float>(s_model.cut_gain) / 1000.0f,
+                                         static_cast<float>(s_model.cut_off) / 1000.0f,
+                                         static_cast<float>(s_model.q_gain) / 1000.0f,
+                                         static_cast<float>(s_model.q_off) / 1000.0f,
+                                         static_cast<float>(s_model.vca_gain) / 1000.0f,
+                                         static_cast<float>(s_model.vca_off) / 1000.0f,
+                                         static_cast<float>(s_model.curve_k) / 100.0f);
 }
 
 void SendCal(uint8_t persist) {
@@ -72,9 +72,9 @@ void SendCal(uint8_t persist) {
 void SendTest() {
     WaveX::Protocol::CvTestMessage test(static_cast<uint8_t>(s_model.group),
                                         s_test_active ? 1 : 0,
-                                        s_model.test_cut / 100.0f,
-                                        s_model.test_res / 100.0f,
-                                        s_model.test_vca / 100.0f);
+                                        static_cast<float>(s_model.test_cut) / 100.0f,
+                                        static_cast<float>(s_model.test_res) / 100.0f,
+                                        static_cast<float>(s_model.test_vca) / 100.0f);
     if (inter_mcu_send_cv_test(test) != ESP_OK) {
         ESP_LOGW(TAG, "CV test send failed");
     }
